Update Slack alert function code
